fix(ignore): recursive ignore rules match should now behave like in git
According to the documentation of ignore, each provided path should be relative to the
folder containing the associated ignore rules.

This commit addresses that, to behave like in git

Moreover, pattern matching is done in reverse order, starting from the closest rules to
the furthest ones. If a rule explicitly uningnores an entry, then that entry is considered
unignored. Otherwise we traverse the whole stack of rules