fix!: ignore behavior now matches that of git (#104)
* fix(ignore): recursive ignore rules match should now behave like in git

According to the documentation of ignore, each provided path should be relative to the
folder containing the associated ignore rules.

This commit addresses that, to behave like in git

Moreover, pattern matching is done in reverse order, starting from the closest rules to
the furthest ones. If a rule explicitly uningnores an entry, then that entry is considered
unignored. Otherwise we traverse the whole stack of rules

* test(ignore): add test on the ignore behavior

* chore: bump version to 0.7.0

* chore: update README

Add note on the newly introduced support for the ingore everything unignore something pattern
Add warning on the zip support (only 32 bit)

* build: update @types/node dev dependency

* build: update @inquirer/confirm dependency

* chore: fix typos